+++ controllers/helper/registry_spec.go
@@ -1,11 +1,12 @@
-package v1alpha1
+package helper
 
 import (
 	legacymodel "github.com/mittwald/goharbor-client/v3/apiv2/model/legacy"
+	"github.com/mittwald/harbor-operator/api/v1alpha1"
 )
 
 // ToHarborRegistry returns a Harbor registry constructed from the provided spec.
-func (spec *RegistrySpec) ToHarborRegistry(id int64, credential *legacymodel.RegistryCredential) *legacymodel.Registry {
+func ToHarborRegistry(spec *v1alpha1.RegistrySpec, id int64, credential *legacymodel.RegistryCredential) *legacymodel.Registry {
 	return &legacymodel.Registry{
 		ID:          id,
 		Name:        spec.Name,
